requirements.txt
Django==2.0.7
django-autocomplete-light==3.2.10
django-taggit==0.22.2
django-wysiwyg-redactor==0.5.1
six==1.11.0